+++ playwright.config.ts
@@ -1,0 +1,55 @@
+import path from 'node:path';
+
+import { defineConfig, devices } from '@playwright/test';
+import 'dotenv/config';
+
+// Reference: https://playwright.dev/docs/test-configuration
+export default defineConfig({
+  // Directory where tests are located
+  testDir: path.join('tests', 'end-to-end'),
+
+  // Maximum time one test can run for
+  timeout: 30 * 1000,
+
+  // Run tests in files in parallel
+  fullyParallel: true,
+
+  // Fail the build on CI if you accidentally left test.only in the source code
+  forbidOnly: !!process.env.CI,
+
+  // Retry failed tests on CI
+  retries: process.env.CI ? 2 : 0,
+
+  // Opt out of parallel tests on CI
+  workers: process.env.CI ? 1 : undefined,
+
+  // Reporter to use
+  reporter: 'html',
+
+  // Shared settings for all the projects below
+  use: {
+    // Base URL to use in actions like `await page.goto('/')`
+    baseURL: 'http://localhost:3020',
+
+    // Collect trace when retrying the failed test
+    trace: 'on-first-retry',
+
+    // Take screenshot on failure
+    screenshot: 'only-on-failure',
+  },
+
+  // Configure projects for different browsers
+  projects: [
+    {
+      name: 'chromium',
+      use: { ...devices['Desktop Chrome'] },
+    },
+  ],
+
+  // Run local dev server before starting the tests
+  webServer: {
+    command: 'npm start',
+    port: 3020,
+    reuseExistingServer: !process.env.CI,
+  },
+});