+++ src/components/content/RescueMeRecyclePromo/RescueMeRecyclePromo.tsx
@@ -1,13 +1,9 @@
 import { CustomElement } from '@/types/customElement';
 
-export interface IconLinkAttributes {
-  border?: boolean;
-}
-
 declare module 'react' {
   namespace JSX {
     interface IntrinsicElements {
-      'locator-icon-link': CustomElement<IconLinkAttributes>;
+      'locator-rescue-me-recycle-promo': CustomElement;
     }
   }
 }