+++ src/lib/getTip.ts
@@ -1,9 +1,13 @@
 import * as Sentry from '@sentry/browser';
 import random from 'lodash/random';
 
+import i18n from '@/lib/i18n';
 import { RecyclingMeta } from '@/types/locatorApi';
 
 import LocatorApi from './LocatorApi';
+
+type Country = 'ENGLAND' | 'WALES';
+let tipCountryPromise: Promise<Country>;
 
 /**
  * Get a tip for a material or path, falling back to a random generic tip
@@ -13,6 +17,7 @@ export default function getTip(
   options: {
     path?: string;
     materialId?: string | number;
+    country?: string;
   } = {},
 ): RecyclingMeta {
   const tips = [];
@@ -43,10 +48,47 @@ function handleTipError(error: Error) {
   });
 }
 
-export async function getTipByPath(path: string) {
+/**
+ * Get the country for the tip
+ * Saves the promise to avoid repeated checks
+ */
+async function getTipCountry(): Promise<Country> {
+  if (tipCountryPromise === undefined) {
+    tipCountryPromise = new Promise((resolve) => {
+      i18n.on('initialized', () => {
+        const isWelshLocale =
+          i18n.language === 'cy' || i18n.language === 'cy-GB';
+        const isWalesRecycles = window.location.host.includes('walesrecycles');
+        resolve(isWelshLocale || isWalesRecycles ? 'WALES' : 'ENGLAND');
+      });
+    });
+  }
+  return tipCountryPromise;
+}
+
+/**
+ * Get a tip for a given path name
+ *
+ * @param path - The path to get the tip for (a path can be set in the admin)
+ * @param fallback – whether to fallback to a random tip if no tip is found for the path
+ */
+export async function getTipByPath(
+  path: string,
+  { fallback }: { fallback?: boolean } = { fallback: true },
+) {
   try {
-    const meta = await LocatorApi.get<RecyclingMeta[]>('recycling-meta');
-    return getTip(meta, { path });
+    const country = await getTipCountry();
+    const url = `recycling-meta?categories=HINT&country=${country}`;
+
+    if (fallback) {
+      const meta = await LocatorApi.get<RecyclingMeta[]>(url);
+
+      return getTip(meta, { path });
+    } else {
+      const tips = await LocatorApi.get<RecyclingMeta[]>(`${url}&path=${path}`);
+
+      return tips?.[0] ?? null;
+    }
   } catch (error) {
     handleTipError(error);
     return Promise.resolve(null);
@@ -55,8 +97,9 @@ export async function getTipByPath(path: string) {
 
 export async function getTipByMaterial(materialId: string) {
   try {
+    const country = await getTipCountry();
     const meta = await LocatorApi.get<RecyclingMeta[]>(
-      'recycling-meta?categories=HINT',
+      `recycling-meta?categories=HINT&country=${country}`,
     );
     return getTip(meta, { materialId });
   } catch (error) {