+++ src/pages/start.routes.tsx
@@ -7,6 +7,7 @@ import startAction, {
   homeRecyclingStartAction,
   materialStartAction,
 } from './start.action';
+import startLoader from './start.loader';
 import StartPage from './start.page';
 
 const routes: RouteObject[] = [
@@ -14,18 +15,21 @@ const routes: RouteObject[] = [
     path: '/',
     element: <StartPage />,
     action: startAction,
+    loader: startLoader,
     errorElement: <NotFoundPage />,
   },
   {
     path: '/home-recycling',
     element: <HomeRecyclingStartPage />,
     action: homeRecyclingStartAction,
+    loader: startLoader,
     errorElement: <NotFoundPage />,
   },
   {
     path: '/material',
     element: <MaterialStartPage />,
     action: materialStartAction,
+    loader: startLoader,
     errorElement: <NotFoundPage />,
   },
 ];