+++ tests/utils/providei18n.ts
@@ -1,12 +1,13 @@
 import i18n from 'i18next';
 
-import en from '../../public/translations/en.json';
+import en from '../../public/translations/en.json' with { type: 'json' };
 
 export default function provideI18n() {
-  return new Promise<void>((resolve) => {
+  return new Promise<typeof i18n>((resolve) => {
     i18n.init(
       {
         lng: 'en',
+        fallbackLng: 'en',
         debug: false,
         ns: ['translations'],
         defaultNS: 'translations',
@@ -16,7 +17,7 @@ export default function provideI18n() {
           },
         },
       },
-      () => resolve(),
+      () => resolve(i18n),
     );
   });
 }